requirements.txt
aiogram==3.12.0
python-dotenv
beautifulsoup4
requests

git+https://github.com/Kiny-Kiny/EromeAPI.git